+++ .lgtm.yml
@@ -9,4 +9,4 @@ path_classifiers:
     - versioneer.py  # Set Versioneer.py to an external "library" (3rd party code)
     - devtools/*
   generated:
-    - covid_moonshot_ml/_version.py
+    - asap_mlmodels/_version.py
