+++ asap_mlmodels/_version.py
@@ -42,7 +42,7 @@ def get_config():
     cfg.style = "pep440"
     cfg.tag_prefix = ""
     cfg.parentdir_prefix = "None"
-    cfg.versionfile_source = "covid_moonshot_ml/_version.py"
+    cfg.versionfile_source = "asap_mlmodels/_version.py"
     cfg.verbose = False
     return cfg
