+++ asap_mlmodels/tests/test_covid_moonshot_ml.py
@@ -1,5 +1,5 @@
 """
-Unit and regression test for the covid_moonshot_ml package.
+Unit and regression test for the asap_mlmodels package.
 """
 
 # Import package, test suite, and other packages as needed
@@ -8,4 +8,4 @@ import sys
 
 def test_covid_moonshot_ml_imported():
     """Sample test, will always pass so long as import statement worked."""
-    assert "covid_moonshot_ml" in sys.modules
+    assert "asap_mlmodels" in sys.modules
