+++ scripts/cdd_to_schema.py
@@ -3,7 +3,7 @@ import os
 import sys
 
 sys.path.append(f"{os.path.dirname(os.path.abspath(__file__))}/../")
-from asap_dataset_handling.datasets.cdd import cdd_to_schema, cdd_to_schema_pair
+from asap_datasets.datasets.cdd import cdd_to_schema, cdd_to_schema_pair
 
 
 def get_args():
